requirements.txt
openai-whisper
torch
torchaudio
numpy
ffmpeg-python
pydub
moviepy
loguru
python-dotenv
requests
transformers
pyannote.audio
pyannote.core
pyannote.metrics
